requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
python-dateutil>=2.8.2
icalendar>=5.0.0
